Estimates the average effective reproductive number

Averages the time-level reproductive numbers within a certain range to estimate the overall reproductive number for an oubreak.

estimateRtAvg(rtData, rangeForAvg = NULL)

Arguments

rtData
The name of the dateset with time-level reproductive numbers.

rangeForAvg
A vector with the start and ending time period to be used to calculate the average effective reproductive number.

Value

A data frame with the average effective reproductive. Column names:

RtAvg - the average time-level reproductive number between the range specified in rangeForAvg.

Details

This function is meant to be called by estimateR which estimates the individual-level and time-level, and average reproductive numbers, but it can also be run directly.

See also